% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitlmerAB.R
\name{fitlmerAB}
\alias{fitlmerAB}
\title{Run mixed-effects model for two-factor within-subject design using (\code{\link[lme4]{lmer}})}
\usage{
fitlmerAB(mcr.data, ri.only = FALSE, wsbi = FALSE, NoCovar = FALSE,
  ncond = 2, LRT = TRUE, krmodcomp = FALSE, pbmodcomp = FALSE)
}
\arguments{
\item{mcr.data}{A dataframe formatted as generated by \code{\link{generate_data}}.}

\item{ri.only}{Whether the random effects specification is to be
random-intercepts only (TRUE) or maximal random-effects (FALSE).}

\item{wsbi}{Whether the design is between-items (TRUE) or within-items
(FALSE).}

\item{NoCovar}{If TRUE, model to be fitted will be a no-covariance model.
If FALSE, model to be fitted will be the maximal model.}

\item{ncond}{Number of levels of the condition variable}

\item{krmodcomp}{If TRUE calculates F-statistics with Kenward-Roger correction using \code{\link[pbkrtest]{KRmodcomp}}}

\item{pbmodcomp}{If TRUE calculates parametrically bootstrapped LRT effects using \code{\link[pbkrtest]{PBmodcomp}}}
}
\value{
A vector with elements:
If LRT = true:
\item{convmain}{Code signifying convergence of main model
for use in LRTs and bootstrapping (REML = FALSE):
(0): model converged without issue
(1) : model converged after upping the iterations to 250,000
(2) : model did not converge after upping the iterations}
\item{convCond}{Code signifying convergence of LRT model without Condition (REML = FALSE):
(0): model converged without issue
(1) : model converged after upping the iterations to 250,000
(2) : model did not converge after upping the iterations}
\item{convMeas}{Code signifying convergence of LRT model without Measurement Time: (REML = FALSE)
(0): model converged without issue
(1) : model converged after upping the iterations to 250,000
(2) : model did not converge after upping the iterations}
\item{convInt}{Code signifying convergence of LRT model without Interaction (REML = FALSE):
(0): model converged without issue
(1) : model converged after upping the iterations to 250,000
(2) : model did not converge after upping the iterations}
\item{t.Cond}{t-statistic for the condition effect}
\item{pt.Cond}{p-value for the t-statistic of condition (normal distribution) }
\item{chi.Cond}{chi-square statistic for the likelihood ratio test (1 df)
for the condition effect }
\item{pchi.Cond}{p-value for the chi-square statistic for condition effect }
\item{t.Meas}{t-statistic for the measurement effect}
\item{pt.Meas}{p-value for the t-statistic of measurement (normal distribution)}
\item{chi.Meas}{chi-square statistic for the likelihood ratio test (1 df) for the measurement effect}
\item{pchi.Meas}{p-value for the chi-square statistic for measurement effect}
\item{t.Int}{t-statistic for the interaction effect }
\item{pt.Int}{p-value for the t-statistic of interaction (normal distribution) }
\item{chi.Int}{chi-square statistic for the likelihood ratio test (1 df)
for the interaction effect }
\item{pchi.Int}{p-value for the chi-square statistic for interaction effect }

If krmodcomp = TRUE:
\item{convmain.kr}{Code signifying convergence of main model
for use with krmodcomp (REML=TRUE):
(0): model converged without issue
(1) : model converged after upping the iterations to 250,000
(2) : model did not converge after upping the iterations}
\item{convCond.kr}{Code signifying convergence of krmodcomp model without Condition (REML=TRUE):
(0): model converged without issue
(1) : model converged after upping the iterations to 250,000
(2) : model did not converge after upping the iterations}
\item{convMeas.kr}{Code signifying convergence of krmodcomp model without Measurement Time (REML=TRUE):
(0): model converged without issue
(1) : model converged after upping the iterations to 250,000
(2) : model did not converge after upping the iterations}
\item{convInt.kr}{Code signifying convergence of krmodcomp model without Interaction (REML=TRUE):
(0): model converged without issue
(1) : model converged after upping the iterations to 250,000
(2) : model did not converge after upping the iterations}
\item{f.Cond}{F-statistic with Kenward-Roger correction for condition effect }
\item{pf.Cond}{p-value for the F-statistic with Kenward-Roger correction for
condition effect}
\item{f.Meas}{F-statistic with Kenward-Roger correction for measurement effect}
\item{pf.Meas}{p-value for the F-statistic with Kenward-Roger correction for
measurement effect}
\item{f.Int}{F-statistic with Kenward-Roger correction for interaction effect}
\item{pf.Int}{p-value for the F-statistic with Kenward-Roger correction for
interaction effect}

If pbmodcomp = TRUE:
\item{PBtest.Cond}{p-value of Condition determined as the fraction of simulated LRT-values that are larger
or equal to the observed LRT value}
\item{Bartlett.Cond}{p-value of Condition determined from the mean of simulated LRT-values}
\item{Gamma.Cond}{p-value of Condition determined from a gamma distribution with mean and variance
determined as the sample mean and variance of simulated LRT-values}
\item{PBtest.Meas}{p-value of Measurement Time determined as the fraction of simulated LRT-values that are larger
or equal to the observed LRT value}
\item{Bartlett.Meas}{p-value of Measurement Time determined from the mean of simulated LRT-values}
\item{Gamma.Meas}{p-value of Measurement Time determined from a gamma distribution with mean and variance
determined as the sample mean and variance of simulated LRT-values}
\item{PBtest.Int}{p-value of Interaction determined as the fraction of simulated LRT-values that are larger
or equal to the observed LRT value}
\item{Bartlett.Int}{p-value of Interaction determined from the mean of simulated LRT-values}
\item{Gamma.Int}{p-value of Interaction determined from a gamma distribution with mean and variance
determined as the sample mean and variance of simulated LRT-values}
}
\description{
Runs either a random-intercepts only model, a maximal
random-effects model (random intercepts and slopes for both subject and item for
within-item designs; random intercepts for both subject and item, but only random
slopes for subject for between-item designs), or a no-covariance random-effect
model (same as maximal model but without covariances between random intercepts
and slopes). This function is especially tailored for use in simulations
using \code{\link[simgen]{mcRun}}.
NB: this function is for two-factor within-subject design with one
variable representing condition and another variable representing measurement time.
}
\details{
\code{fitlmerAB} will attempt to fit the model specified by the user.
\code{fitlmerAB} calculates p-values by
(1) using the t-statistic as an approximation for z-distribution (only when ncond = 2);
(2) likelihood ratio tests, (3) conditional F-tests with Kenward-Roger correction
of degrees-of-freedom through \code{\link[pbkrtest]{KRmodcomp}},
(4) parametric bootstrapping using \code{\link[pbkrtest]{PBmodcomp}}.
}
\examples{
#Design with 4 condition levels
design_args <- list(ivs = c(Cond = 4), n_item = 24, n_rep=2)
pop_params <- gen_pop(design_args, 24)
xd <- sim_norm(design_args, 12, pop_params)
colnames(xd)[colnames(xd)=="n_rep"] <- 'Meas'
colnames(xd)[colnames(xd)=="Y"] <- 'Resp'
colnames(xd)[colnames(xd)=="subj_id"] <- 'SubjID'
colnames(xd)[colnames(xd)=="item_id"] <- 'ItemID'
xd <- with_dev_pred(xd, c("Cond", "Meas"))

# Maximal model for within-item design
fitlmerAB(xd, ncond=4, NoCovar=FALSE)

# Maximal model for between-subject design
fitlmerAB(xd, ncond=4, wsbi=TRUE, NoCovar=FALSE)

#Random intercept only design
fitlmerAB(xd, ncond=4, ri.only=TRUE)

#Maximal within-item model with 4 condition levels for use in mcRun to run 1000 simulations
my_des <- list(ivs = c(Cond = 4), n_item = 24, n_rep = 2)
pops <- replicate(1000, gen_pop(my_des, 24), simplify = FALSE)

result <- mcRun("fitlmerAB",
mcr.fnArgs = list(ncond=4, NoCovar=FALSE)
mcr.datFn = "generate_data",
mcr.constant = list(design_args = my_des, n_subj = 24),
mcr.varying = pops)
}
\seealso{
\code{\link{generate_data}}, \code{\link[funfact]{sim_norm}}, \code{\link[simgen]{mcRun}}
}

