% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitRManova.R
\name{fitRManova}
\alias{fitRManova}
\title{Run repeated measures ANOVA for two-factor within-subject design using \code{\link[ez]{ezANOVA}}}
\usage{
fitRManova(mcr.data)
}
\arguments{
\item{mcr.data}{A dataframe formatted as described in \code{\link{generate_data}}.}
}
\value{
A vector with elements:
\item{f.Cond}{F-statistic for the condition effect}
\item{pf.Cond}{p-value for the F-statistic of condition}
\item{f.Meas}{F-statistic for the measurement time effect}
\item{pf.Meas}{p-value for the F-statistic of measurement time}
\item{f.Int}{F-statistic for the interaction effect}
\item{pf.Int}{p-value for the F-statistic of interaction}
}
\description{
Runs a Repeated Measures ANOVA; NB: this function is for
two-factor within-subject design with one variable representing
condition and another variable representing measurement time.
This function was developed for use in simulations using \code{\link[simgen]{mcRun}}
}
\details{
\code{fitRManova} calculates p-values by using type III F-tests
}
\examples{
#Design with 4 condition levels
design_args <- list(ivs = c(Cond = 4), n_item = 24, n_rep=2)
pop_params <- gen_pop(design_args, 24)
xd <- sim_norm(design_args, 12, pop_params)
colnames(xd)[colnames(xd)=="n_rep"] <- 'Meas'
colnames(xd)[colnames(xd)=="Y"] <- 'Resp'
colnames(xd)[colnames(xd)=="subj_id"] <- 'SubjID'
colnames(xd)[colnames(xd)=="item_id"] <- 'ItemID'
xd <- with_dev_pred(xd, c("Cond", "Meas"))

# run RM Anova
fitRManova(xd)
#RM anova with 4 condition levels for use in mcRun to run 1000 simulations
my_des <- list(ivs = c(Cond = 4), n_item = 24, n_rep = 2)
pops <- replicate(1000, gen_pop(my_des, 24), simplify = FALSE)

result <- mcRun("fitRManova",
mcr.datFn = "generate_data",
mcr.constant = list(design_args = my_des, n_subj = 24),
mcr.varying = pops)
}
\seealso{
\code{\link{generate_data}} \code{\link[simgen]{mcRun}}
}

