% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generates two-factor within-subject design using (\code{\link[funfact]{sim_norm}})}
\usage{
generate_data(mcr.params)
}
\arguments{
\item{mcr.params}{Design arguments  as used by \code{\link[funfact]{sim_norm}}. For details
see \code{\link[funfact]{stim_lists}}.}
}
\description{
Generates two-factor within-subject design with one variable representing condition
and another variable representing measurement time. This function is designed for
generating data for simulations run by \code{\link[simgen]{mcRun}}.
}
\examples{
#Simulation fitting two-factor maximal mixed-effect models
my_des <- list(ivs = c(Cond = 4), n_item = 24, n_rep = 2)
pops <- replicate(1000, gen_pop(my_des, 24), simplify = FALSE)

result <- mcRun("fitlmer.facWithinABMax",
mcr.fnArgs = list(ncond=4)
mcr.datFn = "generate_data",
mcr.constant = list(design_args = my_des, n_subj = 24),
mcr.varying = pops)
}
\seealso{
\code{\link[funfact]{sim_norm}}, \code{\link[simgen]{mcRun}}, \code{\link{fitlmer.facWithinABMax}}, \code{\link{fitlmer.facWithinABNoCovar}}
}

