% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{symptom}
\alias{symptom}
\title{Responses to symptoms from a sample of the general population
of Pakistan.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{id}{participant ID, integer}
  \item{age5}{age in 5 year bins, 
          (18,20) (20,25) (25,30) (30,35) 
          (35,40) (40,45) (45,50) (50,55) 
          (55,60) (60,65)}
  \item{gender}{female, male, character}
  \item{bmq_spec}{Pakistan adaption of Beliefs about Medicines Questionnaire
                  (Specific) Stored as POMP score 0-100}
  \item{bmq_necess}{Pakistan adaption of Beliefs about Medicines Questionnaire
                  (Necessity) Stored as POMP score 0-100}
  \item{bmq_concern}{Pakistan adaption of Beliefs about Medicines Questionnaire,
                  (Concern) Stored as POMP score 0-100}
  \item{bmq_general}{Pakistan adaption of Beliefs about Medicines Questionnaire,
                  (General) Stored as POMP score 0-100}
  \item{bis}{Pakistan adaption of Barratt Impuslivity Scale,
                  Stored as POMP score 0-100}        
  \item{symp_days_reported}{Number of days on which symptoms were reported,
                  Non-negative integer (days)}
  \item{actual_help_days}{Number of days on which participants visited some
                  type of health professional, Non-negative integer}
     }
}
\source{
\url{https://osf.io/4mjhq/}

data from Anwar M, Green JA, Norris P, et al 
Prospective daily diary study reporting of any and all symptoms in healthy 
adults in Pakistan: prevalence and #' response
BMJ Open 2017;7:e014998
\doi{10.1136/bmjopen-2016-014998}
}
\usage{
symptom
}
\description{
A dataset containing the age, gender, number of days on which symptoms
were experienced, number of days on which help was sought, as well as 
measures of impulsivity and attitudes to medicines.
}
\keyword{datasets}
